feat: add SSR-friendly about section and dynamic component split
